requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
playwright==1.40.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
allure-pytest==2.
pytest-order==1.3.0
pytest-asyncio>=0.21.0
